Move socket_handler to plugin_utils.
